Auto stash before merge of "develop" and "origin/develop"

+++ monitors/adblock.js
@@ -12,7 +12,7 @@ module.exports = class extends Monitor {
   }
 
   async run(message) {
-    if (!message.guild || !message.guild.settings.adblock.enabled) return null
+    if (!message.guild || !message.guild.settings.adblock.enabled || !message.guild.settings.adblock.exclude[message.guild.channels.id]) return null
     if (await message.hasAtLeastPermissionLevel(6)) return null
     if (!/(https?:\/\/)?(www\.)?(discord\.(gg|li|me|io)|discordapp\.com\/invite)\/.+/.test(message.content)) return null
     return message.delete().catch(err => this.client.emit('log', err, 'error'))
